Split receiver result-domain tests
